requirements.txt
numpy
pandas
Flask
gunicorn
scikit-learn
pandas
PyYAML
evidently
dill
matplotlib
-e .